+++ adapters/unreal/Source/KnoemaUnreal/KnoemaUnreal.Build.cs
@@ -1,0 +1,29 @@
+using UnrealBuildTool;
+
+public class KnoemaUnreal : ModuleRules
+{
+    public KnoemaUnreal(ReadOnlyTargetRules Target) : base(Target)
+    {
+        PCHUsage = PCHUsageMode.UseExplicitOrSharedPCHs;
+        CppStandard = CppStandardVersion.Cpp20;
+
+        PublicDependencyModuleNames.AddRange(
+            new[]
+            {
+                "Core",
+                "CoreUObject",
+                "Engine",
+                "HTTP",
+                "Json",
+                "JsonUtilities"
+            }
+        );
+
+        PrivateDependencyModuleNames.AddRange(
+            new[]
+            {
+                "Projects"
+            }
+        );
+    }
+}