+++ adapters/unreal/Source/KnoemaUnreal/Private/KnoemaUnrealModule.cpp
@@ -1,0 +1,7 @@
+#include "Modules/ModuleManager.h"
+
+class FKnoemaUnrealModule : public IModuleInterface
+{
+};
+
+IMPLEMENT_MODULE(FKnoemaUnrealModule, KnoemaUnreal)